+++ kmath-commons/src/main/kotlin/space/kscience/kmath/commons/transform/Transformations.kt
@@ -6,7 +6,9 @@ import kotlinx.coroutines.flow.map
 import org.apache.commons.math3.transform.*
 import space.kscience.kmath.complex.Complex
 import space.kscience.kmath.streaming.chunked
+import space.kscience.kmath.streaming.spread
 import space.kscience.kmath.structures.*
+
 
 
 /**