+++ kmath-core/src/commonMain/kotlin/space/kscience/kmath/nd/DoubleFieldND.kt
@@ -17,10 +17,10 @@ public class DoubleFieldND(
     ScaleOperations<StructureND<Double>>,
     ExtendedField<StructureND<Double>> {
 
-    override val zero: NDBuffer<Double> by lazy { produce { zero } }
-    override val one: NDBuffer<Double> by lazy { produce { one } }
+    override val zero: BufferND<Double> by lazy { produce { zero } }
+    override val one: BufferND<Double> by lazy { produce { one } }
 
-    override fun number(value: Number): NDBuffer<Double> {
+    override fun number(value: Number): BufferND<Double> {
         val d = value.toDouble() // minimize conversions
         return produce { d }
     }
@@ -31,31 +31,31 @@ public class DoubleFieldND(
                 this@DoubleFieldND.shape,
                 shape
             )
-            this is NDBuffer && this.strides == this@DoubleFieldND.strides -> this.buffer as DoubleBuffer
+            this is BufferND && this.strides == this@DoubleFieldND.strides -> this.buffer as DoubleBuffer
             else -> DoubleBuffer(strides.linearSize) { offset -> get(strides.index(offset)) }
         }
 
     @Suppress("OVERRIDE_BY_INLINE")
     override inline fun StructureND<Double>.map(
         transform: DoubleField.(Double) -> Double,
-    ): NDBuffer<Double> {
+    ): BufferND<Double> {
         val buffer = DoubleBuffer(strides.linearSize) { offset -> DoubleField.transform(buffer.array[offset]) }
-        return NDBuffer(strides, buffer)
+        return BufferND(strides, buffer)
     }
 
     @Suppress("OVERRIDE_BY_INLINE")
-    override inline fun produce(initializer: DoubleField.(IntArray) -> Double): NDBuffer<Double> {
+    override inline fun produce(initializer: DoubleField.(IntArray) -> Double): BufferND<Double> {
         val array = DoubleArray(strides.linearSize) { offset ->
             val index = strides.index(offset)
             DoubleField.initializer(index)
         }
-        return NDBuffer(strides, DoubleBuffer(array))
+        return BufferND(strides, DoubleBuffer(array))
     }
 
     @Suppress("OVERRIDE_BY_INLINE")
     override inline fun StructureND<Double>.mapIndexed(
         transform: DoubleField.(index: IntArray, Double) -> Double,
-    ): NDBuffer<Double> = NDBuffer(
+    ): BufferND<Double> = BufferND(
         strides,
         buffer = DoubleBuffer(strides.linearSize) { offset ->
             DoubleField.transform(
@@ -69,34 +69,34 @@ public class DoubleFieldND(
         a: StructureND<Double>,
         b: StructureND<Double>,
         transform: DoubleField.(Double, Double) -> Double,
-    ): NDBuffer<Double> {
+    ): BufferND<Double> {
         val buffer = DoubleBuffer(strides.linearSize) { offset ->
             DoubleField.transform(a.buffer.array[offset], b.buffer.array[offset])
         }
-        return NDBuffer(strides, buffer)
+        return BufferND(strides, buffer)
     }
 
     override fun scale(a: StructureND<Double>, value: Double): StructureND<Double> = a.map { it * value }
 
-    override fun power(arg: StructureND<Double>, pow: Number): NDBuffer<Double> = arg.map { power(it, pow) }
+    override fun power(arg: StructureND<Double>, pow: Number): BufferND<Double> = arg.map { power(it, pow) }
 
-    override fun exp(arg: StructureND<Double>): NDBuffer<Double> = arg.map { exp(it) }
+    override fun exp(arg: StructureND<Double>): BufferND<Double> = arg.map { exp(it) }
 
-    override fun ln(arg: StructureND<Double>): NDBuffer<Double> = arg.map { ln(it) }
+    override fun ln(arg: StructureND<Double>): BufferND<Double> = arg.map { ln(it) }
 
-    override fun sin(arg: StructureND<Double>): NDBuffer<Double> = arg.map { sin(it) }
-    override fun cos(arg: StructureND<Double>): NDBuffer<Double> = arg.map { cos(it) }
-    override fun tan(arg: StructureND<Double>): NDBuffer<Double> = arg.map { tan(it) }
-    override fun asin(arg: StructureND<Double>): NDBuffer<Double> = arg.map { asin(it) }
-    override fun acos(arg: StructureND<Double>): NDBuffer<Double> = arg.map { acos(it) }
-    override fun atan(arg: StructureND<Double>): NDBuffer<Double> = arg.map { atan(it) }
+    override fun sin(arg: StructureND<Double>): BufferND<Double> = arg.map { sin(it) }
+    override fun cos(arg: StructureND<Double>): BufferND<Double> = arg.map { cos(it) }
+    override fun tan(arg: StructureND<Double>): BufferND<Double> = arg.map { tan(it) }
+    override fun asin(arg: StructureND<Double>): BufferND<Double> = arg.map { asin(it) }
+    override fun acos(arg: StructureND<Double>): BufferND<Double> = arg.map { acos(it) }
+    override fun atan(arg: StructureND<Double>): BufferND<Double> = arg.map { atan(it) }
 
-    override fun sinh(arg: StructureND<Double>): NDBuffer<Double> = arg.map { sinh(it) }
-    override fun cosh(arg: StructureND<Double>): NDBuffer<Double> = arg.map { cosh(it) }
-    override fun tanh(arg: StructureND<Double>): NDBuffer<Double> = arg.map { tanh(it) }
-    override fun asinh(arg: StructureND<Double>): NDBuffer<Double> = arg.map { asinh(it) }
-    override fun acosh(arg: StructureND<Double>): NDBuffer<Double> = arg.map { acosh(it) }
-    override fun atanh(arg: StructureND<Double>): NDBuffer<Double> = arg.map { atanh(it) }
+    override fun sinh(arg: StructureND<Double>): BufferND<Double> = arg.map { sinh(it) }
+    override fun cosh(arg: StructureND<Double>): BufferND<Double> = arg.map { cosh(it) }
+    override fun tanh(arg: StructureND<Double>): BufferND<Double> = arg.map { tanh(it) }
+    override fun asinh(arg: StructureND<Double>): BufferND<Double> = arg.map { asinh(it) }
+    override fun acosh(arg: StructureND<Double>): BufferND<Double> = arg.map { acosh(it) }
+    override fun atanh(arg: StructureND<Double>): BufferND<Double> = arg.map { atanh(it) }
 }
 
 public fun AlgebraND.Companion.real(vararg shape: Int): DoubleFieldND = DoubleFieldND(shape)