+++ kmath-core/src/commonMain/kotlin/space/kscience/kmath/nd/ShortRingND.kt
@@ -14,10 +14,10 @@ public class ShortRingND(
 ) : BufferedRingND<Short, ShortRing>(shape, ShortRing, Buffer.Companion::auto),
     NumbersAddOperations<StructureND<Short>> {
 
-    override val zero: NDBuffer<Short> by lazy { produce { zero } }
-    override val one: NDBuffer<Short> by lazy { produce { one } }
+    override val zero: BufferND<Short> by lazy { produce { zero } }
+    override val one: BufferND<Short> by lazy { produce { one } }
 
-    override fun number(value: Number): NDBuffer<Short> {
+    override fun number(value: Number): BufferND<Short> {
         val d = value.toShort() // minimize conversions
         return produce { d }
     }
@@ -26,8 +26,8 @@ public class ShortRingND(
 /**
  * Fast element production using function inlining.
  */
-public inline fun BufferedRingND<Short, ShortRing>.produceInline(crossinline initializer: ShortRing.(Int) -> Short): NDBuffer<Short> {
-    return NDBuffer(strides, ShortBuffer(ShortArray(strides.linearSize) { offset -> ShortRing.initializer(offset) }))
+public inline fun BufferedRingND<Short, ShortRing>.produceInline(crossinline initializer: ShortRing.(Int) -> Short): BufferND<Short> {
+    return BufferND(strides, ShortBuffer(ShortArray(strides.linearSize) { offset -> ShortRing.initializer(offset) }))
 }
 
 public inline fun <R> ShortRing.nd(vararg shape: Int, action: ShortRingND.() -> R): R {